Resolved errors with analytics command

+++ code/get_analysis.py
@@ -15,9 +15,14 @@ def viewOverallBudget(chat_id, bot):
             cat_budget = helper.getCategoryBudgetByCategory(chat_id,cat)
             if cat_budget != '0':
                 category_budget[cat] = cat_budget
-    graphing.overall_split(category_budget)
-    bot.send_photo(chat_id, photo=open("overall_split.png", 'rb'), reply_markup=types.ReplyKeyboardRemove())
-    os.remove("overall_split.png")
+    print(category_budget,"category budget")
+    if category_budget == {}:
+        bot.send_message(chat_id,"You are yet to set your budget for different categories.")
+    else:
+        graphing.overall_split(category_budget)
+        bot.send_photo(chat_id, photo=open("overall_split.png", 'rb'), reply_markup=types.ReplyKeyboardRemove())
+        os.remove("overall_split.png")
+        
 
 def viewSpendWise(chat_id, bot):
     category_spend = {}
@@ -27,7 +32,7 @@ def viewSpendWise(chat_id, bot):
             category_spend[cat] = spend
 
     if category_spend == {}:
-        bot.send_message(chat_id, "No category spend available", reply_markup=types.ReplyKeyboardRemove())
+        bot.send_message(chat_id, "No expenditure available for this month", reply_markup=types.ReplyKeyboardRemove())
         return
     graphing.spend_wise_split(category_spend)
     bot.send_photo(chat_id, photo=open("spend_wise.png", 'rb'), reply_markup=types.ReplyKeyboardRemove())
@@ -42,10 +47,14 @@ def viewRemaining(chat_id, bot):
     for cat in categories:
         if helper.isCategoryBudgetByCategoryAvailable(chat_id, cat):
             percent = helper.calculateRemainingCateogryBudgetPercent(chat_id, cat)
-            category_spend_percent[cat] = percent
-    graphing.remaining(category_spend_percent)
-    bot.send_photo(chat_id, photo=open("remaining.png", 'rb'), reply_markup=types.ReplyKeyboardRemove())
-    os.remove("remaining.png")
+            if percent:
+                category_spend_percent[cat] = percent
+    if category_spend_percent != {}:
+        graphing.remaining(category_spend_percent)
+        bot.send_photo(chat_id, photo=open("remaining.png", 'rb'), reply_markup=types.ReplyKeyboardRemove())
+        os.remove("remaining.png")
+    else:
+        bot.send_message(chat_id,"You are yet to set your budget for different categories.")
 
 def viewHistory(chat_id, bot):
     if not helper.getUserHistory(chat_id):
